# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/classes": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Get all classes",
                "description": "Retrieve a list of classes from the database and return it as a JSON response.",
                "operationId": "getAllClasses",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Class Name"
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "CL1"
                                            },
                                            "start_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-03-01"
                                            },
                                            "end_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-03-01"
                                            },
                                            "schedule": {
                                                "type": "string",
                                                "example": "Monday"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Class schedule at 8:00 AM"
                                            },
                                            "course_id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "courses": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "course_name": {
                                                        "type": "string",
                                                        "example": "Course Name"
                                                    },
                                                    "course_code": {
                                                        "type": "string",
                                                        "example": "Course code"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-classe": {
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Create a new classe",
                "description": "Creates a new classe resource.",
                "operationId": "cb5342be78168f566a4e7e7b252f5bd2",
                "requestBody": {
                    "description": "Classe object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "start_date",
                                    "end_date",
                                    "schedule",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the class",
                                        "type": "string",
                                        "example": "Class A"
                                    },
                                    "code": {
                                        "description": "Code of the class",
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "start_date": {
                                        "description": "Start date of the class",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-24"
                                    },
                                    "end_date": {
                                        "description": "End date of the class",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-30"
                                    },
                                    "course_id": {
                                        "description": "Course id",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "schedule": {
                                        "description": "Class schedule",
                                        "type": "string",
                                        "example": "Monday, Wednesday, Friday"
                                    },
                                    "description": {
                                        "description": "Description of the class",
                                        "type": "string",
                                        "example": "This is Class A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Classe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Classe created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "code": [
                                                    "The code field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-classe/{id}": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Retrieve a specific classe by ID",
                "description": "Retrieves details of a classe based on its ID.",
                "operationId": "d55e0f9eb45d91d7a4a0c351cf091d65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the classe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID of the class",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "description": "Name of the class",
                                            "type": "string",
                                            "example": "Class A"
                                        },
                                        "code": {
                                            "description": "Code of the class",
                                            "type": "string",
                                            "example": "001"
                                        },
                                        "start_date": {
                                            "description": "Start date of the class",
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-03-24"
                                        },
                                        "end_date": {
                                            "description": "End date of the class",
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-06-30"
                                        },
                                        "course_id": {
                                            "description": "Course id",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "schedule": {
                                            "description": "Class schedule",
                                            "type": "string",
                                            "example": "Monday, Wednesday, Friday"
                                        },
                                        "description": {
                                            "description": "Description of the class",
                                            "type": "string",
                                            "example": "This is Class A"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Class not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/delete-classe/{id}": {
            "delete": {
                "tags": [
                    "Classes"
                ],
                "summary": "Delete a specific classe by ID",
                "description": "Deletes a classe based on its ID.",
                "operationId": "e77a85f958db1c2e96e0995ea02b2755",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the classe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Classe deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Classe deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Class not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}